Use cypress for rank tests
